requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
PyMySQL==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
bcrypt==4.1.1
alembic==1.13.1
redis==5.0.1
celery==5.3.4
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
selenium==4.15.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.3.0
jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
cryptography==41.0.8
email-validator==2.1.0
httpx==0.25.2